moved additional widgets + added demo folder for upcoming demos
