+++ crates/rmf_site_editor/src/widgets/view_layers.rs
@@ -28,7 +28,7 @@ use crate::{
 use bevy::{ecs::system::SystemParam, prelude::*};
 use bevy_egui::egui::{Button, CollapsingHeader, DragValue, ScrollArea, Ui};
 use rmf_site_picking::Selection;
-use rmf_site_ui::{Tile, WidgetSystem};
+use rmf_site_ui::{PropertiesTilePlugin, Tile, WidgetSystem};
 
 /// Add a widget for viewing a list of layers
 #[derive(Default)]
