+++ crates/rmf_site_ui/src/lib.rs
@@ -1,16 +1,8 @@
-pub mod menu_bar;
 use bevy_ecs::{prelude::*, system::{SystemParam, SystemState}};
 use bevy_egui::egui::Ui;
-pub use menu_bar::*;
 
-pub mod header_panel;
-pub use header_panel::*;
-
-pub mod panel_of_tiles;
-pub use panel_of_tiles::*;
-
-pub mod panel;
-pub use panel::*;
+pub mod widgets;
+pub use widgets::*;
 
 /// Implement this on a [`SystemParam`] struct to make it a widget that can be
 /// plugged into the site editor UI.
